+++ ta4j-core/src/main/java/org/ta4j/core/num/DoubleNum.java
@@ -37,9 +37,9 @@ public class DoubleNum implements Num {
 
     private static final long serialVersionUID = -2611177221813615070L;
 
-    public static final DoubleNum ZERO = DoubleNum.valueOf(0);
-    public static final DoubleNum ONE = DoubleNum.valueOf(1);
-    public static final DoubleNum HUNDRED = DoubleNum.valueOf(100);
+    private static final DoubleNum ZERO = DoubleNum.valueOf(0);
+    private static final DoubleNum ONE = DoubleNum.valueOf(1);
+    private static final DoubleNum HUNDRED = DoubleNum.valueOf(100);
 
     private final static double EPS = 0.00001; // precision
     private final double delegate;
@@ -70,6 +70,21 @@ public class DoubleNum implements Num {
 
     public static DoubleNum valueOf(Number i) {
         return new DoubleNum(Double.parseDouble(i.toString()));
+    }
+    
+    @Override
+    public Num zero() {
+        return ZERO;
+    }
+
+    @Override
+    public Num one() {
+        return ONE;
+    }
+
+    @Override
+    public Num hundred() {
+        return HUNDRED;
     }
 
     @Override